Crate genesis_token_canister

Statics

GTC

Functions

balance
Returns the token balance of a given address

balance_

canister_init

canister_post_upgrade

canister_pre_upgrade

claim_neurons
Claim the caller's GTC neurons (on behalf of the caller) and return the IDs of these neurons

claim_neurons_

donate_account
Donate the caller's GTC account funds to the Neuron given by the GTC's donate_account_recipient_neuron_id.

donate_account_

forward_all_unclaimed_accounts
Transfer the funds of all unclaimed accounts to the Neuron given by the GTC's forward_all_unclaimed_accounts_recipient_neuron_id.

forward_all_unclaimed_accounts_

get_account
Return the account state of the given GTC address

get_account_

gtc

gtc_mut

len
Returns the number of unique addresses present in the ledger

len_

main

total
Returns the sum of all token balances in the internal ledger

total_